tests/ambient-light/ambient-light_stop.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: stop</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Xiuqi Jiang" href="mailto:xiuqix.jiang@intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<link rel="help" href="https://w3c.github.io/sensors/releases/WD-generic-sensor-20160830.html#dom-sensor-stop">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var stop_result;
let sensor = new AmbientLightSensor();

sensor.start();

stop_result = sensor.stop();

test(function() {
  assert_equals(sensor.state, "idle");
}, "the sensor.state is idle after fire stop event");

test(function() {
  assert_equals(sensor.reading, "null");
}, "the sensor.reading is null after fire stop event");

test(function() {
  try{
    sensor.stop();
  } catch (e){
    assert_equals(e.name, "InvalidStateError");
  }
}, "throw an InvalidStateError exception when state is either idle or errored");

test(function() {
  assert_equals(stop_result, undefined);
}, "the sensor.stop return undefined");

</script>
